One textured-quad pipeline instead of two copies
The image and text renderers each wrote out a full wgpu pipeline. With the
debug labels neutralised the two were identical down to the byte: same
shader file, same vertex format, same premultiplied blend state, same
clamped bilinear sampler, same two triangles.

They now share one. What stays separate is what genuinely differs — how a
texture gets *produced*: one decodes or rasterises an image source, the
other lays glyphs into an atlas, and neither is close to the other.

The point is less the 100 lines than what the duplication allowed: the blend
state could drift on one side and nothing would say so.

image_quad 888 -> 780, text_quad 622 -> 513, plus 137 shared.
