+++ src/renderer/image_quad.rs
@@ -9,14 +9,15 @@ use std::sync::Arc;
 
 use wgpu::util::DeviceExt;
 use wgpu::{
-    BindGroup, BindGroupLayout, Buffer as WgpuBuffer, Device, Extent3d, Queue, RenderPass,
-    RenderPipeline, Sampler, Texture, TextureDimension, TextureFormat, TextureUsages,
+    BindGroup, Buffer as WgpuBuffer, Device, Extent3d, Queue, RenderPass, Texture,
+    TextureDimension, TextureFormat, TextureUsages,
 };
 
 use super::commands::DrawCommand;
 use super::constants::{IMAGE_HASH_SAMPLE_SIZE, SVG_QUALITY_MULTIPLIER};
 use super::flatten::FlattenedCommand;
 use super::gpu::NO_CLIP_RECT;
+use super::textured_quad::TexturedQuadPipeline;
 use super::textured_vertex::{TexturedVertex, to_ndc};
 use crate::widgets::Rect;
 use crate::widgets::image::{ContentFit, ImageSource};
@@ -69,13 +70,9 @@ impl Hash for CacheKey {
 
 /// Renderer for images as textured quads.
 pub struct ImageQuadRenderer {
-    // Quad rendering pipeline
-    pipeline: RenderPipeline,
-    bind_group_layout: BindGroupLayout,
-    sampler: Sampler,
-
-    // Shared index buffer (vertices are per-quad)
-    index_buffer: WgpuBuffer,
+    /// Pipeline, layout, sampler and index buffer — shared with the text
+    /// renderer, which draws the same quad from a different texture.
+    quad: TexturedQuadPipeline,
 
     // Texture cache
     texture_cache: HashMap<CacheKey, Arc<CachedTexture>>,
@@ -89,113 +86,8 @@ pub struct ImageQuadRenderer {
 
 impl ImageQuadRenderer {
     pub fn new(device: &Device, format: TextureFormat) -> Self {
-        // Load shader from dedicated file
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: Some("ImageQuad Shader"),
-            source: wgpu::ShaderSource::Wgsl(include_str!("textured_quad_shader.wgsl").into()),
-        });
-
-        // Create texture bind group layout
-        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            label: Some("ImageQuad Bind Group Layout"),
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Texture {
-                        multisampled: false,
-                        view_dimension: wgpu::TextureViewDimension::D2,
-                        sample_type: wgpu::TextureSampleType::Float { filterable: true },
-                    },
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Filtering),
-                    count: None,
-                },
-            ],
-        });
-
-        // Create pipeline layout
-        let pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-            label: Some("ImageQuad Pipeline Layout"),
-            bind_group_layouts: &[&bind_group_layout],
-            immediate_size: 0,
-        });
-
-        // Create render pipeline
-        let pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: Some("ImageQuad Pipeline"),
-            layout: Some(&pipeline_layout),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: Some("vs_main"),
-                buffers: &[TexturedVertex::desc()],
-                compilation_options: wgpu::PipelineCompilationOptions::default(),
-            },
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: Some("fs_main"),
-                targets: &[Some(wgpu::ColorTargetState {
-                    format,
-                    blend: Some(wgpu::BlendState {
-                        color: wgpu::BlendComponent {
-                            src_factor: wgpu::BlendFactor::SrcAlpha,
-                            dst_factor: wgpu::BlendFactor::OneMinusSrcAlpha,
-                            operation: wgpu::BlendOperation::Add,
-                        },
-                        alpha: wgpu::BlendComponent {
-                            src_factor: wgpu::BlendFactor::One,
-                            dst_factor: wgpu::BlendFactor::OneMinusSrcAlpha,
-                            operation: wgpu::BlendOperation::Add,
-                        },
-                    }),
-                    write_mask: wgpu::ColorWrites::ALL,
-                })],
-                compilation_options: wgpu::PipelineCompilationOptions::default(),
-            }),
-            primitive: wgpu::PrimitiveState {
-                topology: wgpu::PrimitiveTopology::TriangleList,
-                strip_index_format: None,
-                front_face: wgpu::FrontFace::Ccw,
-                cull_mode: None,
-                polygon_mode: wgpu::PolygonMode::Fill,
-                unclipped_depth: false,
-                conservative: false,
-            },
-            depth_stencil: None,
-            multisample: wgpu::MultisampleState::default(),
-            multiview_mask: None,
-            cache: None,
-        });
-
-        // Create sampler
-        let sampler = device.create_sampler(&wgpu::SamplerDescriptor {
-            label: Some("ImageQuad Sampler"),
-            address_mode_u: wgpu::AddressMode::ClampToEdge,
-            address_mode_v: wgpu::AddressMode::ClampToEdge,
-            address_mode_w: wgpu::AddressMode::ClampToEdge,
-            mag_filter: wgpu::FilterMode::Linear,
-            min_filter: wgpu::FilterMode::Linear,
-            mipmap_filter: wgpu::MipmapFilterMode::Nearest,
-            ..Default::default()
-        });
-
-        // Create index buffer
-        let indices: [u16; 6] = [0, 1, 2, 1, 3, 2];
-        let index_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("ImageQuad Index Buffer"),
-            contents: bytemuck::cast_slice(&indices),
-            usage: wgpu::BufferUsages::INDEX,
-        });
-
         Self {
-            pipeline,
-            bind_group_layout,
-            sampler,
-            index_buffer,
+            quad: TexturedQuadPipeline::new(device, format, "ImageQuad"),
             texture_cache: HashMap::new(),
             current_frame: 0,
             max_cache_size: 64,
@@ -650,7 +542,7 @@ impl ImageQuadRenderer {
         // Create bind group
         let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
             label: Some("ImageQuad Bind Group"),
-            layout: &self.bind_group_layout,
+            layout: &self.quad.bind_group_layout,
             entries: &[
                 wgpu::BindGroupEntry {
                     binding: 0,
@@ -658,7 +550,7 @@ impl ImageQuadRenderer {
                 },
                 wgpu::BindGroupEntry {
                     binding: 1,
-                    resource: wgpu::BindingResource::Sampler(&self.sampler),
+                    resource: wgpu::BindingResource::Sampler(&self.quad.sampler),
                 },
             ],
         });
@@ -876,8 +768,8 @@ impl ImageQuadRenderer {
             return;
         }
 
-        render_pass.set_pipeline(&self.pipeline);
-        render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
+        render_pass.set_pipeline(&self.quad.pipeline);
+        render_pass.set_index_buffer(self.quad.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
 
         for quad in quads {
             render_pass.set_bind_group(0, &quad.bind_group, &[]);
